sail_ui: move QtPage to sail_ui
