+++ packages/drivechain_client/lib/pages/overview_page.dart
@@ -5,12 +5,12 @@ import 'package:drivechain_client/providers/blockchain_provider.dart';
 import 'package:drivechain_client/widgets/error_container.dart';
 import 'package:drivechain_client/widgets/qt_container.dart';
 import 'package:drivechain_client/widgets/qt_icon_button.dart';
-import 'package:drivechain_client/widgets/qt_page.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:get_it/get_it.dart';
 import 'package:money2/money2.dart';
 import 'package:sail_ui/sail_ui.dart';
+import 'package:sail_ui/widgets/containers/qt_page.dart';
 import 'package:stacked/stacked.dart';
 
 @RoutePage()
