+++ packages/drivechain_client/lib/pages/sidechain_activation_management_page.dart
@@ -4,11 +4,11 @@ import 'package:drivechain_client/pages/sidechain_proposal_page.dart';
 import 'package:drivechain_client/providers/sidechain_provider.dart';
 import 'package:drivechain_client/widgets/qt_button.dart';
 import 'package:drivechain_client/widgets/qt_icon_button.dart';
-import 'package:drivechain_client/widgets/qt_page.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:get_it/get_it.dart';
 import 'package:sail_ui/sail_ui.dart';
+import 'package:sail_ui/widgets/containers/qt_page.dart';
 import 'package:stacked/stacked.dart';
 
 @RoutePage()
