+++ packages/drivechain_client/lib/pages/sidechain_proposal_page.dart
@@ -1,13 +1,13 @@
-import 'package:drivechain_client/api.dart';
-import 'package:drivechain_client/widgets/qt_icon_button.dart';
-import 'package:drivechain_client/widgets/qt_page.dart';
-import 'package:flutter/material.dart';
 import 'package:auto_route/annotations.dart';
+import 'package:drivechain_client/api.dart';
+import 'package:drivechain_client/widgets/qt_button.dart';
+import 'package:drivechain_client/widgets/qt_container.dart';
+import 'package:drivechain_client/widgets/qt_icon_button.dart';
+import 'package:flutter/material.dart';
 import 'package:get_it/get_it.dart';
 import 'package:sail_ui/sail_ui.dart';
+import 'package:sail_ui/widgets/containers/qt_page.dart';
 import 'package:stacked/stacked.dart';
-import 'package:drivechain_client/widgets/qt_container.dart';
-import 'package:drivechain_client/widgets/qt_button.dart';
 
 @RoutePage()
 class SidechainProposalPage extends StatelessWidget {
